+++ src/elog/inc/elog_queued_target.h
@@ -26,7 +26,7 @@ public:
      * @brief Construct a new ELogQueuedTarget object.
      * @param logTarget The deferred log target.
      */
-    ELogQueuedTarget(ELogTarget* logTarget, uint32_t batchSize, uint32_t timeoutMillis);
+    ELogQueuedTarget(ELogTarget* logTarget, uint32_t batchSize, uint64_t timeoutMillis);
     ELogQueuedTarget(const ELogQueuedTarget&) = delete;
     ELogQueuedTarget(ELogQueuedTarget&&) = delete;
     ELogQueuedTarget& operator=(const ELogQueuedTarget&) = delete;
