+++ src/elog/src/elog_pgsql_db_target_provider.h
@@ -19,7 +19,7 @@ public:
 
     ELogDbTarget* loadTarget(const ELogConfigMapNode* logTargetCfg, const std::string& connString,
                              const std::string& insertQuery, ELogDbTarget::ThreadModel threadModel,
-                             uint32_t maxThreads, uint32_t reconnectTimeoutMillis) final;
+                             uint32_t maxThreads, uint64_t reconnectTimeoutMillis) final;
 };
 
 }  // namespace elog
